optimize: separate speaker from tokenizer
- move `spk_stat.pt` into config
- move `sample_audio_speaker` to DVAE
- update rvcmd to `v0.2.7`
