+++ ChatTTS/model/__init__.py
@@ -1,4 +1,5 @@
 from .dvae import DVAE
 from .gpt import GPT
 from .processors import gen_logits
+from .speaker import Speaker
 from .tokenizer import Tokenizer
