+++ ChatTTS/utils/dl.py
@@ -51,7 +51,6 @@ def check_all_assets(base_dir: Path, sha256_map: Dict[str, str], update=False) -
         "Decoder.pt",
         "DVAE_full.pt",
         "GPT.pt",
-        "spk_stat.pt",
         "Vocos.pt",
     ]
     for model in names:
@@ -114,7 +113,7 @@ def download_dns_yaml(url: str, folder: str):
         logger.get_logger().info(f"downloaded into {folder}")
 
 
-def download_all_assets(tmpdir: str, version="0.2.6"):
+def download_all_assets(tmpdir: str, version="0.2.7"):
     import subprocess
     import platform
