+++ tools/checksum/tmpl.go
@@ -4,7 +4,6 @@ var files = [...]string{
 	"asset/Decoder.pt",
 	"asset/DVAE_full.pt",
 	"asset/GPT.pt",
-	"asset/spk_stat.pt",
 	"asset/Vocos.pt",
 
 	"asset/tokenizer/special_tokens_map.json",
@@ -16,7 +15,6 @@ const jsontmpl = `{
 	"sha256_asset_Decoder_pt"   : "%s",
 	"sha256_asset_DVAE_full_pt"      : "%s",
 	"sha256_asset_GPT_pt"       : "%s",
-	"sha256_asset_spk_stat_pt"  : "%s",
 	"sha256_asset_Vocos_pt"     : "%s",
 
 	"sha256_asset_tokenizer_special_tokens_map_json": "%s",
